Move RoutePath where every generator that needs it can see it
RoutePath landed under Web/Routing/, next to the route tree it was written for.
Two of its five callers - OpenApiDocumentGenerator and LinkGenerator - are
compiled into Hardened.Function.SourceGenerator through <Compile Include> links,
and that project takes Configuration, DependencyInjection, Models, Module,
OpenApiDocument, Requests, Shared and Function. Not Web. So the function
generator stopped compiling.

Shared/ is the set every consumer already links, and a path join is not web
machinery, so that is where it belongs.

Worth recording why this was not caught by the commit that introduced it:
Hardened.Function.SourceGenerator is in the solution but no test project
references it, so `dotnet test` never builds it. Only `dotnet build` does, and
the runs verifying that commit were tests alone. A packable generator that the
test command does not even compile is its own gap; this fixes the breakage, not
the gap.
